fix : Ai 추천 호출 순서 변경
feat
- fix : Ai 추천 호출 순서 변경
